deleting image after changing in edit product or deleting product

+++ controllers/admin.js
@@ -1,4 +1,5 @@
 const Product = require('../models/product')
+const fileHelper = require('../utils/file');
 
 const { validationResult } = require('express-validator');
 
@@ -129,6 +130,7 @@ exports.postEditProduct = (req, res, next) => {
             product.price = updatedPrice;
             product.description = updatedDesc;
             if (image) {
+                fileHelper.deleteFile(product.imageUrl);
                 product.imageUrl = image.path;
             }
             return product.save()
@@ -159,7 +161,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeteleProd = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
+    Product.findById(prodId)
+        .then(product => {
+            if (!product) {
+                return
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({ _id: prodId, userId: req.user._id })
+        })
         .then(result => {
             console.log('product deleted');
             res.redirect('/admin/products');
